Comment allusers+allevents and adapter classes
