--- app/src/main/java/com/example/pygmyhippo/BaseViewHolder.java
@@ -1,14 +1,0 @@
-package com.example.pygmyhippo;
-
-import android.view.View;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-public abstract class BaseViewHolder<T> extends RecyclerView.ViewHolder {
-    public BaseViewHolder(@NonNull View itemView) {
-        super(itemView);
-    }
-
-    public abstract void setViews(T dataclass);
-}
