--- app/src/main/java/com/example/pygmyhippo/RecyclerClickListener.java
@@ -1,5 +1,0 @@
-package com.example.pygmyhippo;
-
-public interface RecyclerClickListener {
-    void onItemClick(int position);
-}
