+++ app/src/main/res/navigation/user_mobile_navigation.xml
@@ -8,14 +8,14 @@
 
     <fragment
         android:id="@+id/navigation_all_events"
-        android:name="com.example.pygmyhippo.AllEventsFragment"
+        android:name="com.example.pygmyhippo.admin.AllEventsFragment"
         android:label="All List"
         tools:layout="@layout/fragment_all_list"
         android:background="@color/dark_purple"/>
 
     <fragment
         android:id="@+id/navigation_all_users"
-        android:name="com.example.pygmyhippo.AllUsersFragment"
+        android:name="com.example.pygmyhippo.admin.AllUsersFragment"
         android:label="All List"
         tools:layout="@layout/fragment_all_list"
         android:background="@color/dark_purple"/>
